feat: support immutable interfaced target types
