+++ src/Riok.Mapperly/Descriptors/WellKnownTypes.cs
@@ -39,12 +39,16 @@ public class WellKnownTypes
     private INamedTypeSymbol? _immutableArrayT;
     private INamedTypeSymbol? _immutableList;
     private INamedTypeSymbol? _immutableListT;
+    private INamedTypeSymbol? _iImmutableListT;
     private INamedTypeSymbol? _immutableHashSet;
     private INamedTypeSymbol? _immutableHashSetT;
+    private INamedTypeSymbol? _iImmutableSetT;
     private INamedTypeSymbol? _immutableQueue;
     private INamedTypeSymbol? _immutableQueueT;
+    private INamedTypeSymbol? _iImmutableQueueT;
     private INamedTypeSymbol? _immutableStack;
     private INamedTypeSymbol? _immutableStackT;
+    private INamedTypeSymbol? _iImmutableStackT;
     private INamedTypeSymbol? _immutableSortedSet;
     private INamedTypeSymbol? _immutableSortedSetT;
     private INamedTypeSymbol? _immutableDictionary;
@@ -90,12 +94,16 @@ public class WellKnownTypes
     public INamedTypeSymbol ImmutableArrayT => _immutableArrayT ??= GetTypeSymbol(typeof(ImmutableArray<>));
     public INamedTypeSymbol ImmutableList => _immutableList ??= GetTypeSymbol(typeof(ImmutableList));
     public INamedTypeSymbol ImmutableListT => _immutableListT ??= GetTypeSymbol(typeof(ImmutableList<>));
+    public INamedTypeSymbol IImmutableListT => _iImmutableListT ??= GetTypeSymbol(typeof(IImmutableList<>));
     public INamedTypeSymbol ImmutableHashSet => _immutableHashSet ??= GetTypeSymbol(typeof(ImmutableHashSet));
     public INamedTypeSymbol ImmutableHashSetT => _immutableHashSetT ??= GetTypeSymbol(typeof(ImmutableHashSet<>));
+    public INamedTypeSymbol IImmutableSetT => _iImmutableSetT ??= GetTypeSymbol(typeof(IImmutableSet<>));
     public INamedTypeSymbol ImmutableQueue => _immutableQueue ??= GetTypeSymbol(typeof(ImmutableQueue));
     public INamedTypeSymbol ImmutableQueueT => _immutableQueueT ??= GetTypeSymbol(typeof(ImmutableQueue<>));
+    public INamedTypeSymbol IImmutableQueueT => _iImmutableQueueT ??= GetTypeSymbol(typeof(IImmutableQueue<>));
     public INamedTypeSymbol ImmutableStack => _immutableStack ??= GetTypeSymbol(typeof(ImmutableStack));
     public INamedTypeSymbol ImmutableStackT => _immutableStackT ??= GetTypeSymbol(typeof(ImmutableStack<>));
+    public INamedTypeSymbol IImmutableStackT => _iImmutableStackT ??= GetTypeSymbol(typeof(IImmutableStack<>));
     public INamedTypeSymbol ImmutableSortedSet => _immutableSortedSet ??= GetTypeSymbol(typeof(ImmutableSortedSet));
     public INamedTypeSymbol ImmutableSortedSetT => _immutableSortedSetT ??= GetTypeSymbol(typeof(ImmutableSortedSet<>));
     public INamedTypeSymbol ImmutableDictionary => _immutableDictionary ??= GetTypeSymbol(typeof(ImmutableDictionary));
